# update wasm-opt arguments with more optimization passes