--- a/embed/build.sh
+++ b/embed/build.sh
@@ -26,7 +26,23 @@ trap 'rm -f sqlite3.tmp' EXIT
 	$(awk '{print "-Wl,--export="$0}' exports.txt)
 
 "$BINARYEN/wasm-ctor-eval" -g -c _initialize sqlite3.wasm -o sqlite3.tmp
-"$BINARYEN/wasm-opt" \
+
+# For more information on arguments passed to
+# wasm-opt please see `wasm-opt --help` and:
+# https://github.com/WebAssembly/binaryen/wiki/Optimizer-Cookbook
+#
+# --debuginfo            : emit "names" section (useful during Go panics)
+# --optimize-level 4     : set code optimization level to max
+# --shrink-level 4       : set code shrinking level to max
+# --strip-producers      : strip the wasm "producers" section
+# --dce                  : dead code elimination
+# --vacuum               : remove more obviously un-needed code
+# --type-ssa             : creates new nominal types to help optimizations
+# -Oz                    : a combined set of optimization passes focused on *size*
+# --type-merging         : merges to get rid of above used ssa types
+# -O3                    : a combined set of optimization passes focused on *speed*
+# --converge             : re-runs the whole set of passes while binary size keeps shrinking
+"$BINARYEN/wasm-opt" --debuginfo \
 	sqlite3.tmp -o sqlite3.wasm \
 	--enable-simd --enable-mutable-globals --enable-multivalue \
 	--enable-bulk-memory --enable-reference-types \
@@ -35,12 +51,8 @@ trap 'rm -f sqlite3.tmp' EXIT
 	--shrink-level 4 \
 	--strip-producers \
 	--dce --vacuum \
-	--flatten \
-	--rereloop \
-	-Oz -Oz \
-	--gufa-optimizing -O4 \
-	--inlining-optimizing \
-	--precompute-propagate \
-	--optimize-instructions \
-	--inlining-optimizing \
+	--type-ssa \
+	-Oz \
+	--type-merging \
+	-O3 \
 	--converge