Suggestion for some improved wasm-opt arguments. It has a slight increase in binary size, but also an increase in performance running a number of the benchmarks.

Running the stats benchmarks before:
```
goos: linux
goarch: amd64
pkg: github.com/ncruces/go-sqlite3/ext/stats
cpu: AMD Ryzen 7 7840U w/ Radeon  780M Graphics     
Benchmark_average
Benchmark_average-16     	19713591	        61.12 ns/op	       0 B/op	       0 allocs/op
Benchmark_average-16     	19514550	        60.92 ns/op	       0 B/op	       0 allocs/op
Benchmark_average-16     	19225606	        61.85 ns/op	       0 B/op	       0 allocs/op
Benchmark_average-16     	18870000	        62.72 ns/op	       0 B/op	       0 allocs/op
Benchmark_average-16     	18717904	        61.35 ns/op	       0 B/op	       0 allocs/op
Benchmark_average-16     	18917599	        60.92 ns/op	       0 B/op	       0 allocs/op
Benchmark_average-16     	18866498	        60.92 ns/op	       0 B/op	       0 allocs/op
Benchmark_average-16     	18855162	        61.51 ns/op	       0 B/op	       0 allocs/op
Benchmark_average-16     	18696751	        64.41 ns/op	       0 B/op	       0 allocs/op
Benchmark_average-16     	18590013	        62.31 ns/op	       0 B/op	       0 allocs/op
Benchmark_variance
Benchmark_variance-16    	 7961498	       148.9 ns/op	       0 B/op	       0 allocs/op
Benchmark_variance-16    	 7860943	       147.3 ns/op	       0 B/op	       0 allocs/op
Benchmark_variance-16    	 7896700	       149.7 ns/op	       0 B/op	       0 allocs/op
Benchmark_variance-16    	 8040048	       148.4 ns/op	       0 B/op	       0 allocs/op
Benchmark_variance-16    	 7862599	       148.4 ns/op	       0 B/op	       0 allocs/op
Benchmark_variance-16    	 8055547	       147.2 ns/op	       0 B/op	       0 allocs/op
Benchmark_variance-16    	 8138631	       148.3 ns/op	       0 B/op	       0 allocs/op
Benchmark_variance-16    	 7935478	       149.5 ns/op	       0 B/op	       0 allocs/op
Benchmark_variance-16    	 7797823	       148.5 ns/op	       0 B/op	       0 allocs/op
Benchmark_variance-16    	 8003864	       149.5 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/ncruces/go-sqlite3/ext/stats	26.382s
```

Running the stats benchmarks after:
```
goos: linux
goarch: amd64
pkg: github.com/ncruces/go-sqlite3/ext/stats
cpu: AMD Ryzen 7 7840U w/ Radeon  780M Graphics     
Benchmark_average
Benchmark_average-16     	20117323	        58.64 ns/op	       0 B/op	       0 allocs/op
Benchmark_average-16     	20225779	        58.81 ns/op	       0 B/op	       0 allocs/op
Benchmark_average-16     	20348558	        58.91 ns/op	       0 B/op	       0 allocs/op
Benchmark_average-16     	20003413	        58.50 ns/op	       0 B/op	       0 allocs/op
Benchmark_average-16     	20216776	        59.25 ns/op	       0 B/op	       0 allocs/op
Benchmark_average-16     	19537520	        58.98 ns/op	       0 B/op	       0 allocs/op
Benchmark_average-16     	20088400	        61.03 ns/op	       0 B/op	       0 allocs/op
Benchmark_average-16     	20155567	        60.31 ns/op	       0 B/op	       0 allocs/op
Benchmark_average-16     	19393212	        59.57 ns/op	       0 B/op	       0 allocs/op
Benchmark_average-16     	20177713	        59.50 ns/op	       0 B/op	       0 allocs/op
Benchmark_variance
Benchmark_variance-16    	 7834611	       142.0 ns/op	       0 B/op	       0 allocs/op
Benchmark_variance-16    	 8137812	       143.9 ns/op	       0 B/op	       0 allocs/op
Benchmark_variance-16    	 8119240	       148.4 ns/op	       0 B/op	       0 allocs/op
Benchmark_variance-16    	 8201199	       144.0 ns/op	       0 B/op	       0 allocs/op
Benchmark_variance-16    	 8245419	       145.8 ns/op	       0 B/op	       0 allocs/op
Benchmark_variance-16    	 8101179	       147.7 ns/op	       0 B/op	       0 allocs/op
Benchmark_variance-16    	 8216662	       150.5 ns/op	       0 B/op	       0 allocs/op
Benchmark_variance-16    	 8187937	       145.1 ns/op	       0 B/op	       0 allocs/op
Benchmark_variance-16    	 8070690	       146.2 ns/op	       0 B/op	       0 allocs/op
Benchmark_variance-16    	 8082541	       146.3 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/ncruces/go-sqlite3/ext/stats	26.522s
```

Running the parallel benchmarks before:
```
goos: linux
goarch: amd64
pkg: github.com/ncruces/go-sqlite3/tests/parallel
cpu: AMD Ryzen 7 7840U w/ Radeon  780M Graphics     
Benchmark_parallel
Benchmark_parallel-16    	    5392	    311109 ns/op	  289730 B/op	     718 allocs/op
Benchmark_parallel-16    	    5036	    310859 ns/op	  289970 B/op	     718 allocs/op
Benchmark_parallel-16    	    3756	    298261 ns/op	  289891 B/op	     718 allocs/op
Benchmark_parallel-16    	    3649	    297470 ns/op	  289871 B/op	     718 allocs/op
Benchmark_parallel-16    	    5522	    308925 ns/op	  290090 B/op	     718 allocs/op
Benchmark_parallel-16    	    3894	    302616 ns/op	  289705 B/op	     717 allocs/op
Benchmark_parallel-16    	    5042	    310545 ns/op	  289781 B/op	     718 allocs/op
Benchmark_parallel-16    	    4293	    298550 ns/op	  289627 B/op	     718 allocs/op
Benchmark_parallel-16    	    4540	    300154 ns/op	  289852 B/op	     718 allocs/op
Benchmark_parallel-16    	    4779	    303570 ns/op	  289759 B/op	     718 allocs/op
Benchmark_wal
Benchmark_wal-16         	    2704	    393413 ns/op	  292100 B/op	     692 allocs/op
Benchmark_wal-16         	    2936	    395608 ns/op	  292055 B/op	     691 allocs/op
Benchmark_wal-16         	    2654	    402413 ns/op	  292100 B/op	     692 allocs/op
Benchmark_wal-16         	    2618	    395084 ns/op	  292081 B/op	     692 allocs/op
Benchmark_wal-16         	    2895	    393415 ns/op	  292114 B/op	     691 allocs/op
Benchmark_wal-16         	    2860	    395407 ns/op	  292104 B/op	     691 allocs/op
Benchmark_wal-16         	    2894	    402127 ns/op	  292125 B/op	     692 allocs/op
Benchmark_wal-16         	    2714	    396850 ns/op	  292107 B/op	     692 allocs/op
Benchmark_wal-16         	    2763	    398753 ns/op	  292086 B/op	     692 allocs/op
Benchmark_wal-16         	    2763	    398274 ns/op	  292111 B/op	     692 allocs/op
Benchmark_memdb
Benchmark_memdb-16       	    6757	    337622 ns/op	  230913 B/op	     620 allocs/op
Benchmark_memdb-16       	    4770	    318074 ns/op	  230917 B/op	     620 allocs/op
Benchmark_memdb-16       	    6067	    328663 ns/op	  230923 B/op	     620 allocs/op
Benchmark_memdb-16       	    4725	    319037 ns/op	  230924 B/op	     620 allocs/op
Benchmark_memdb-16       	    4544	    314241 ns/op	  230921 B/op	     620 allocs/op
Benchmark_memdb-16       	    4182	    311591 ns/op	  230934 B/op	     620 allocs/op
Benchmark_memdb-16       	    7581	    348758 ns/op	  230906 B/op	     620 allocs/op
Benchmark_memdb-16       	    3512	    303049 ns/op	  230931 B/op	     620 allocs/op
Benchmark_memdb-16       	    4118	    311353 ns/op	  230936 B/op	     620 allocs/op
Benchmark_memdb-16       	    4730	    312024 ns/op	  230932 B/op	     620 allocs/op
PASS
ok  	github.com/ncruces/go-sqlite3/tests/parallel	44.094s
```

Running the parallel benchmarks after:
```
goos: linux
goarch: amd64
pkg: github.com/ncruces/go-sqlite3/tests/parallel
cpu: AMD Ryzen 7 7840U w/ Radeon  780M Graphics     
Benchmark_parallel
Benchmark_parallel-16    	    4298	    288136 ns/op	  292019 B/op	     717 allocs/op
Benchmark_parallel-16    	    5120	    291389 ns/op	  292251 B/op	     717 allocs/op
Benchmark_parallel-16    	    4460	    285547 ns/op	  291994 B/op	     717 allocs/op
Benchmark_parallel-16    	    4849	    294373 ns/op	  292127 B/op	     717 allocs/op
Benchmark_parallel-16    	    4566	    299726 ns/op	  292030 B/op	     717 allocs/op
Benchmark_parallel-16    	    3616	    282404 ns/op	  291967 B/op	     717 allocs/op
Benchmark_parallel-16    	    4488	    294051 ns/op	  292233 B/op	     717 allocs/op
Benchmark_parallel-16    	    4365	    288414 ns/op	  292132 B/op	     717 allocs/op
Benchmark_parallel-16    	    4264	    274768 ns/op	  292021 B/op	     717 allocs/op
Benchmark_parallel-16    	    4304	    287452 ns/op	  292141 B/op	     717 allocs/op
Benchmark_wal
Benchmark_wal-16         	    2960	    380842 ns/op	  294287 B/op	     691 allocs/op
Benchmark_wal-16         	    3058	    387471 ns/op	  294324 B/op	     691 allocs/op
Benchmark_wal-16         	    2860	    389543 ns/op	  294342 B/op	     691 allocs/op
Benchmark_wal-16         	    2751	    390249 ns/op	  294315 B/op	     691 allocs/op
Benchmark_wal-16         	    2762	    386735 ns/op	  294331 B/op	     691 allocs/op
Benchmark_wal-16         	    3249	    396590 ns/op	  294317 B/op	     691 allocs/op
Benchmark_wal-16         	    2925	    389004 ns/op	  294341 B/op	     691 allocs/op
Benchmark_wal-16         	    2980	    386498 ns/op	  294332 B/op	     691 allocs/op
Benchmark_wal-16         	    3064	    389224 ns/op	  294318 B/op	     691 allocs/op
Benchmark_wal-16         	    2751	    384150 ns/op	  294340 B/op	     691 allocs/op
Benchmark_memdb
Benchmark_memdb-16       	    5307	    311794 ns/op	  233200 B/op	     619 allocs/op
Benchmark_memdb-16       	    4898	    302492 ns/op	  233204 B/op	     619 allocs/op
Benchmark_memdb-16       	    6062	    333801 ns/op	  233196 B/op	     619 allocs/op
Benchmark_memdb-16       	    3637	    295519 ns/op	  233223 B/op	     619 allocs/op
Benchmark_memdb-16       	    4096	    291428 ns/op	  233213 B/op	     619 allocs/op
Benchmark_memdb-16       	    8644	    344074 ns/op	  233185 B/op	     619 allocs/op
Benchmark_memdb-16       	    4944	    314662 ns/op	  233192 B/op	     619 allocs/op
Benchmark_memdb-16       	    4016	    303091 ns/op	  233217 B/op	     620 allocs/op
Benchmark_memdb-16       	    4731	    313436 ns/op	  233208 B/op	     619 allocs/op
Benchmark_memdb-16       	    4404	    313063 ns/op	  233215 B/op	     619 allocs/op
PASS
ok  	github.com/ncruces/go-sqlite3/tests/parallel	42.893s
```