@ncruces 2024-10-03
Few points:
- longer compilation times don't bother me much (though they're long as it is)
- maintaining magic flags I don't really understand (rather than generic -On ones) worries me more
- isn't this dropping debugging info?

The thing I'd really like to keep is function names in native panics. It's bad enough that, with aggressive inlining it becomes harder to figure out what happened in a crash.