@NyaaaWhatsUpDoc 2024-10-04
> Few points:
> 
>     * longer compilation times don't bother me much (though they're long as it is)
> 
>     * maintaining magic flags I don't really understand (rather than generic -On ones) worries me more
> 
>     * isn't this dropping debugging info?
> 
> 
> The thing I'd really like to keep is function names in native panics. It's bad enough that, with aggressive inlining it becomes harder to figure out what happened in a crash.

Yep that makes sense about the debug info! Though I think currently you might still be stripping that out. Your current set of args `-g --strip` essentially means `--debuginfo --strip-debug` so I think it ultimately gets removed :p

Regarding magic flags, would it help if I added code comments with explanations for each of them?