@ncruces 2024-10-04
> Yep that makes sense about the debug info! Though I think currently you might still be stripping that out. Your current set of args `-g --strip` essentially means `--debuginfo --strip-debug` so I think it ultimately gets removed :p

The goal is to strip line numbers, and allow all manner of aggressive inlining, but not mangle ("optimize") function names.
This means I still get minimally readable stack traces when something breaks.