@NyaaaWhatsUpDoc 2024-10-04
while trying to confirm that my changes would maintain symbol names, i have unfortunately run into a bit of a problem. running `.github/workflows/repro.sh` to setup necessary toolchains, then running `embed/build.sh` even on current `main` branch causes a change in the `sqlite3.wasm` file. it seems whatever it going on with my own setup *always* seems to strip out symbol names, oddly.

at least until i figure that out so i can more accurately confirm it maintains all the information it needs to, i'm going to close this PR :thinking: