@ncruces 2024-10-04
To check if something strips symbols, the easiest way is to just panic inside a callback.

For example, add a panic in the `CollationNeeded` callback:
https://github.com/ncruces/go-sqlite3/blob/ac94a5406e65e275abfedd1bcbde1952b11f109b/func_test.go#L32-L37

And run the test:
```
--- FAIL: ExampleConn_CreateCollation (0.00s)
panic: test (recovered by wazero)
        wasm stack trace:
                env.go_collation_needed(i32,i32,i32,i32)
                sqlite3.wasm.sqlite3GetCollSeq(i32,i32,i32,i32) i32
                sqlite3.wasm.sqlite3ExprCollSeq(i32,i32) i32
                sqlite3.wasm.sqlite3KeyInfoFromExprList(i32,i32,i32) i32
                sqlite3.wasm.sqlite3Select(i32,i32,i32) i32
                sqlite3.wasm.sqlite3RunParser(i32,i32) i32
                sqlite3.wasm.sqlite3Prepare(i32,i32,i32,i32,i32,i32,i32) i32
                sqlite3.wasm.sqlite3LockAndPrepare(i32,i32,i32,i32,i32,i32,i32) i32
                sqlite3.wasm.sqlite3_prepare_v3(i32,i32,i32,i32,i32,i32) i32 [recovered]
        panic: test (recovered by wazero)
        wasm stack trace:
                env.go_collation_needed(i32,i32,i32,i32)
                sqlite3.wasm.sqlite3GetCollSeq(i32,i32,i32,i32) i32
                sqlite3.wasm.sqlite3ExprCollSeq(i32,i32) i32
                sqlite3.wasm.sqlite3KeyInfoFromExprList(i32,i32,i32) i32
                sqlite3.wasm.sqlite3Select(i32,i32,i32) i32
                sqlite3.wasm.sqlite3RunParser(i32,i32) i32
                sqlite3.wasm.sqlite3Prepare(i32,i32,i32,i32,i32,i32,i32) i32
                sqlite3.wasm.sqlite3LockAndPrepare(i32,i32,i32,i32,i32,i32,i32) i32
                sqlite3.wasm.sqlite3_prepare_v3(i32,i32,i32,i32,i32,i32) i32

goroutine 1 [running]:
testing.(*InternalExample).processRunResult(0xc0036a7c08, {0x0, 0x0}, 0x1cf9732?, 0x0, {0x1fc8280, 0xc000d4bee0})
        /usr/local/go/src/testing/example.go:91 +0x645
testing.runExample.func2()
        /usr/local/go/src/testing/run_example.go:59 +0x116
panic({0x1fc8280?, 0xc000d4bee0?})
        /usr/local/go/src/runtime/panic.go:785 +0x132
github.com/ncruces/go-sqlite3.(*sqlite).call(0xc002344708, {0x1f26233, 0x12?}, {0xc0036a79c8?, 0xc002344a30?, 0x9?})
        /Users/cruces/Documents/Personal/go-sqlite3/sqlite.go:181 +0x1ad
github.com/ncruces/go-sqlite3.(*Conn).PrepareFlags(0xc002a42180, {0x1f339dc, 0x32}, 0x8?)
        /Users/cruces/Documents/Personal/go-sqlite3/conn.go:191 +0x1b1
github.com/ncruces/go-sqlite3.(*Conn).Prepare(...)
        /Users/cruces/Documents/Personal/go-sqlite3/conn.go:172
github.com/ncruces/go-sqlite3_test.ExampleConn_CreateCollation()
        /Users/cruces/Documents/Personal/go-sqlite3/func_test.go:43 +0x186
testing.runExample({{0x1f2ac1c, 0x1b}, 0x2026c70, {0x1f2f6ee, 0x25}, 0x0})
        /usr/local/go/src/testing/run_example.go:63 +0x2de
testing.runExamples(0x202f990?, {0x236bee0, 0x7, 0xf?})
        /usr/local/go/src/testing/example.go:40 +0x126
testing.(*M).Run(0xc000110460)
        /usr/local/go/src/testing/testing.go:2036 +0x71c
main.main()
        _testmain.go:89 +0x9b
FAIL    github.com/ncruces/go-sqlite3   1.510s
FAIL
```

Not amazing, but still useful.